# Optimize Hubris and Regexp