--- a/mruby-gems/src/rubygems/rack_protection.rs
+++ b/mruby-gems/src/rubygems/rack_protection.rs
@@ -30,7 +30,7 @@ impl RackProtection {
                 .map_err(|_| MrbError::SourceNotFound(path.to_owned()))?;
             string = string.replace(
                 "define_method(:default_options) { super().merge(options) }",
-                "@default_options ||= DEFAULT_OPTIONS; @default_options.merge(options)",
+                "define_method(:default_options) { @default_options ||= DEFAULT_OPTIONS; @default_options.merge(options) }",
             );
             string = string.replace(
                 "def default_options\n        DEFAULT_OPTIONS\n      end",