--- a/mruby/src/extn/core/object.rb
+++ b/mruby/src/extn/core/object.rb
@@ -1,0 +1,25 @@
+# frozen_string_literal: true
+
+class NilClass
+  def dup
+    self
+  end
+end
+
+class TrueClass
+  def dup
+    self
+  end
+end
+
+class FalseClass
+  def dup
+    self
+  end
+end
+
+class Integer
+  def dup
+    self
+  end
+end