--- a/mruby/src/extn/core/object.rb
+++ b/mruby/src/extn/core/object.rb
@@ -23,3 +23,9 @@ class Integer
     self
   end
 end
+
+class Symbol
+  def dup
+    self
+  end
+end