Hubris + Sinatra turns out to be one giant integration test for `Regexp`. This PR improves the throughput of hubris by a factor of 24 and reduces p99 latency by a factor of 17.

# Optimizations

## Class

- Memoize `RClass` pointers on class specs.

## Kernel

- Implement `Warning` in Rust.
- Implement `Kernel#warn` in Rust.
- Short circuit `MrbWarn::warn` if `$stderr` is nil.

## Regexp

- Only set and unset the required number of `Regexp` capture group globals, e.g. `$1`, `$2`, etc.
- Implement `Regexp::compile` directly with `Regexp::initialize` without going through the mruby VM.
- Avoid recompiling onig Regex by storing it on the `Regexp` Rust object.
- Implement `Regexp::escape` directly with `mrb_obj_new` without going through the mruby VM.
- Implement Options strings with a truth table and `&'static str`.
- Implement Encoding strings with a truth table and `&'static str`.
- Hand roll the argspec as a constant in `mod regexp::initialize`.

# Nemesis

- Do not GC in `ExecMode::finalize`. mruby will GC on its own.

# Hubris

- Only define `EchoServer` class on interpreter init.
- Bump `max_requests` from 150 to 500.

# Benches

Run hubris in release mode:

```sh
cargo run --release --bin hubris
```

## master

```sh
ab -c 8 -n 10000 -l "http://127.0.0.1:8000/ferrocarril?boom"
```

Elapsed: 259s
Throughput: 39 req/s
p50 latency: 199ms
p99 latency: 309ms

## this PR @ `a2aea45`

```sh
ab -c 8 -n 10000 -l "http://127.0.0.1:8000/ferrocarril?boom"
```

Elapsed: 10.4s
Throughput: 947 req/s
p50 latency: 7ms
p99 latency: 18ms